pyproject.toml
[project]
name = "dispachai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "dj-database-url>=3.0.1",
    "django-rest-knox>=5.0.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
]
